refactored to better utils spot
