+++ src/applications/check-in/utils/api/api.utils.unit.spec.js
@@ -1,0 +1,11 @@
+// import { expect } from 'chai';
+
+// import { makeApiCall } from './utils';
+
+describe('check in', () => {
+  describe('api utils', () => {
+    describe('makeApiCall', () => {
+      it('', () => {});
+    });
+  });
+});
